// web/src/api/generated/models/GiftIdeaDetail.ts
export interface GiftIdeaDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof GiftIdeaDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof GiftIdeaDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    description: string | null;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    occasion: string | null;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    ownerId: string;
    /**
     * 
     * @type {boolean}
     * @memberof GiftIdeaDetail
     */
    pinned: boolean;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    priceText: string | null;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    recipient: string | null;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    spaceId: string;
    /**
     * 
     * @type {GiftIdeaStatus}
     * @memberof GiftIdeaDetail
     */
    status: GiftIdeaStatus;
    /**
     * 
     * @type {Date}
     * @memberof GiftIdeaDetail
     */
    targetOn: Date | null;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof GiftIdeaDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof GiftIdeaDetail
     */
    url: string | null;
    /**
     * 
     * @type {number}
     * @memberof GiftIdeaDetail
     */
    version: number;
}